fix: jest errors, move search route to engines
