+++ x-pack/plugins/enterprise_search/server/routes/enterprise_search/engines.test.ts
@@ -231,4 +231,71 @@ describe('engines routes', () => {
       mockRouter.shouldThrow(request);
     });
   });
+
+  describe('GET /internal/enterprise_search/engines/{engine_name}/search', () => {
+    let mockRouter: MockRouter;
+
+    beforeEach(() => {
+      jest.clearAllMocks();
+      mockRouter = new MockRouter({
+        method: 'get',
+        path: '/internal/enterprise_search/engines/{engine_name}/search',
+      });
+
+      registerEnginesRoutes({
+        ...mockDependencies,
+        router: mockRouter.router,
+      });
+    });
+    it('creates a request to enterprise search', () => {
+      expect(mockRequestHandler.createRequest).toHaveBeenCalledWith({
+        path: '/api/engines/:engine_name/_search',
+      });
+    });
+
+    it('validates correctly with engine_name and pagination', () => {
+      const request = {
+        body: {
+          query: 'test-query',
+          fields: ['test-field-1', 'test-field-2'],
+        },
+        params: {
+          engine_name: 'some-engine',
+          from: 0,
+          size: 10,
+        },
+      };
+
+      mockRouter.shouldValidate(request);
+    });
+
+    it('validates correctly with default pagination and no body', () => {
+      const request = {
+        params: {
+          engine_name: 'my-test-engine',
+        },
+      };
+
+      mockRouter.shouldValidate(request);
+    });
+
+    it('validation with query and without fields', () => {
+      const request = {
+        params: {
+          engine_name: 'my-test-engine',
+        },
+        body: {
+          query: 'sample-query',
+          fields: [],
+        },
+      };
+      mockRouter.shouldValidate(request);
+    });
+
+    it('fails validation without engine name', () => {
+      const request = { params: {} };
+
+      mockRouter.shouldThrow(request);
+    });
+  });
 });
